server/requirements.txt
Flask==2.0.1
google-cloud-bigquery==2.34.4
google-cloud-storage==2.7.0
pandas==1.5.3
gunicorn==20.1.0
flask-cors==3.0.10
google-auth==2.16.2
matplotlib==3.7.1
numpy==1.24.2 